codegen file with types added

+++ frontend/src/components/Templates/TaggedParties/index.tsx
@@ -3,27 +3,20 @@ import { useTag } from "../../../hooks/useTag";
 import { useParams } from "react-router-dom";
 import { useAuth } from "../../../hooks/useAuth";
 import { MultiLevelList } from "../../List";
-import { Typography, Box, Button } from "@mui/material";
+import { Typography, Box } from "@mui/material";
 import { useTranslation } from "react-i18next";
-import { ModalContext } from "../../../contexts/ModalContext";
 import { StyledPaper } from "../../Container";
+import { ExtendedTagParty } from "../../../types/codegen";
 
 export const TaggedParties = () => {
   const { id: recordId } = useParams(); //always string
   const { user } = useAuth();
   const { t } = useTranslation();
-  //const {handleModal} = React.useContext(ModalContext);
-
-  // const toggleNewTagPartyModal = () => {
-  //     handleModal("NewTagParty");
-  //   };
-
   const { operations } = useTag();
 
   const [getTaggedPartiesHandler, getTaggedPartiesRequest] =
     operations.getTaggedParties;
-  const [deleteTagPartyHandler, deleteTagPartyRequest] =
-    operations.deleteTagParty;
+  const [deleteTagPartyHandler] = operations.deleteTagParty;
 
   const untag = (partyId: string) => {
     deleteTagPartyHandler({
@@ -44,41 +37,44 @@ export const TaggedParties = () => {
     });
   }, []);
 
+  if (getTaggedPartiesRequest.loading) {
+    return <div>loading...</div>;
+  }
+
+  const taggedParties: ExtendedTagParty[] =
+    getTaggedPartiesRequest?.data && getTaggedPartiesRequest.data.taggedParties;
+
   return (
     <>
-      {getTaggedPartiesRequest.data?.taggedParties && (
-        <StyledPaper
-          sx={{
-            width: {
-              xs: "100%", // theme.breakpoints.up('xs')
-              sm: "60%", //400, // theme.breakpoints.up('sm')
-              // md: 300, // theme.breakpoints.up('md')
-              lg: "360px", // theme.breakpoints.up('lg')
-              //xl: 500, // theme.breakpoints.up('xl')
-            },
-            marginBottom: {
-                xs: 2,
-                lg: 0
-              }
-          }}
-        >
-          <Box>
-            <Typography variant="h6">
-              {t("singleRecord.connections")}
-            </Typography>
+      <StyledPaper
+        sx={{
+          width: {
+            xs: "100%", // theme.breakpoints.up('xs')
+            sm: "60%", //400, // theme.breakpoints.up('sm')
+            // md: 300, // theme.breakpoints.up('md')
+            lg: "360px", // theme.breakpoints.up('lg')
+            //xl: 500, // theme.breakpoints.up('xl')
+          },
+          marginBottom: {
+            xs: 2,
+            lg: 0,
+          },
+        }}
+      >
+        <Box>
+          <Typography variant="h6">{t("singleRecord.connections")}</Typography>
 
-            {getTaggedPartiesRequest.data.taggedParties.length > 0 ? (
-              <MultiLevelList
-                listName="taggedParties"
-                data={getTaggedPartiesRequest.data.taggedParties}
-                deleteItem={untag}
-              />
-            ) : (
-              <Typography mt={2}>{t("general.noRecords")}</Typography>
-            )}
-          </Box>
-        </StyledPaper>
-      )}
+          {taggedParties?.length > 0 ? (
+            <MultiLevelList
+              listName="taggedParties"
+              data={taggedParties}
+              deleteItem={untag}
+            />
+          ) : (
+            <Typography mt={2}>{t("general.noRecords")}</Typography>
+          )}
+        </Box>
+      </StyledPaper>
     </>
   );
 };
